requirements.txt
fastapi==0.110.0
uvicorn==0.22.0
SQLAlchemy==2.0.44
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-auth==2.22.0
requests==2.32.0
email-validator==2.1.0.post1
python-multipart==0.0.9
psycopg2-binary==2.9.9
fastapi-cors==0.0.6
Jinja2==3.1.2
playwright==1.40.0
